test: Refactor test helpers and update imports
