+++ src/commands/add.rs
@@ -20,7 +20,7 @@ pub fn git_add(path_str: &str) -> Result<(), git2::Error> {
 #[cfg(test)]
 mod tests {
     use crate::commands;
-    use crate::tests::get_repo;
+    use crate::test_helpers::get_repo;
     use serial_test::serial;
     use std::fs::File;
     use std::path::Path;
