+++ src/commands/init.rs
@@ -8,13 +8,14 @@ pub fn git_init() -> Result<(), git2::Error> {
 
 #[cfg(test)]
 mod tests {
+    use crate::test_helpers::get_repo;
     use serial_test::serial;
     use std::path::Path;
 
     #[test]
     #[serial]
     fn test_git_init() {
-        let _repo = crate::tests::get_repo();
+        let _repo = get_repo();
         assert!(Path::new(".git").exists());
     }
 }
