+++ src/commands/log.rs
@@ -19,7 +19,7 @@ pub fn git_log() -> Result<Vec<String>, git2::Error> {
 #[cfg(test)]
 mod tests {
     use crate::commands;
-    use crate::tests::get_repo;
+    use crate::test_helpers::get_repo;
     use serial_test::serial;
     use std::fs::File;
     use std::path::Path;
