+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+#[cfg(test)]
+pub mod test_helpers;
